refacto navigation
