+++ mobile/App.tsx
@@ -1,18 +1,9 @@
 import React, { useState, useEffect } from "react";
-import { NavigationContainer } from "@react-navigation/native";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 
-import HomeScreen from "./src/screens/HomeScreen";
-import SearchScreen from "./src/screens/SearchScreen";
-import AccountScreen from "./src/screens/AccountScreen";
-import iconForRoute from "./src/components/IconForRoute";
-import tw from "./lib/tailwind";
-import LoadedFont from './src/utils/LoadedFont';
+import LoadedFont from "./src/utils/LoadedFont";
+import Navigation from "./src/components/Navigation";
 
-import Ionicons from "@expo/vector-icons/Ionicons";
 import * as SplashScreen from "expo-splash-screen";
-
-const Tab = createBottomTabNavigator();
 
 SplashScreen.preventAutoHideAsync();
 
@@ -30,31 +21,11 @@ export default function App() {
         setAppIsReady(true);
       }
     }
-
     prepare();
   }, []);
 
   if (!appIsReady) {
     return null;
   }
-  return (
-    <NavigationContainer>
-      <Tab.Navigator
-        screenOptions={({ route }) => ({
-          tabBarIcon: ({ focused, color, size }) => {
-            color = focused
-              ? tw.color("tertiary") || color
-              : tw.color("primary") || color;
-            let icon = iconForRoute(route, size);
-            return <Ionicons name={icon.name} size={icon.size} color={color} />;
-          },
-          tabBarShowLabel: false,
-        })}
-      >
-        <Tab.Screen name="Home" component={HomeScreen} initialParams={{appIsReady:appIsReady}}/>
-        <Tab.Screen name="Account" component={AccountScreen} />
-        <Tab.Screen name="Search" component={SearchScreen} />
-      </Tab.Navigator>
-    </NavigationContainer>
-  );
+  return <Navigation appIsReady={appIsReady} />;
 }
